Add panel
Add table panels

Tab height

Table height

Bump project version
