+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-# AgTable.jl Changelog
+# AgTables.jl Changelog
 
-The latest version of this file can be found at the master branch of the [AgTable.jl repository](https://github.com/bhftbootcamp/AgTable.jl).
+The latest version of this file can be found at the master branch of the [AgTables.jl repository](https://github.com/bhftbootcamp/AgTables.jl).
