+++ docs/Project.toml
@@ -1,3 +1,3 @@
 [deps]
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
-AgTable = "de8a86c3-a64b-47c8-8ca2-357579be7910"
+AgTables = "de8a86c3-a64b-47c8-8ca2-357579be7910"
