+++ frontend/public/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>AgTable</title>
+    <title>AgTables</title>
     <style>
         html,
         body {
