+++ test/Project.toml
@@ -3,4 +3,4 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 Serde = "db9b398d-9517-45f8-9a95-92af99003e0e"
 UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
-AgTable = "ade2ca70-3891-5945-98fb-dc099432e06c"
+AgTables = "ade2ca70-3891-5945-98fb-dc099432e06c"
