+++ test/runtests.jl
@@ -4,14 +4,14 @@ using Test
 using Dates
 using UUIDs
 using Serde
-using AgTable
+using AgTables
 
 const skip_fileds = Set{Symbol}([
     :uuid_key,
     :license_key,
 ])
 
-function Base.:(==)(l::T, r::T) where {T<:Union{<:AGTable,<:AbstractColumnDef}}
+function Base.:(==)(l::T, r::T) where {T<:Union{<:AGTable,<:AbstractColumnDef,<:AGPanel}}
     for field in fieldnames(typeof(l))
         if field ∉ skip_fileds
             l_field = getfield(l, field)
@@ -30,7 +30,7 @@ const tests = [
     "common",
 ]
 
-@testset "AgTable" begin
+@testset "AgTables" begin
     @info("Running tests:")
 
     for test ∈ tests
